refactor(ai-agent): split factory::build_agent per agent (Stage 8)
factory/mod.rs was 548 lines with a single 275-line match expression
handling all five agent types. This PR splits each arm into its own
sibling file (acp.rs, aionrs.rs, openclaw.rs, nanobot.rs, remote.rs)
with a shared FactoryContext in context.rs for workspace resolution.

build_agent is now a 10-arm dispatcher ≤ 60 lines; each agent's build
logic lives in its own ≤ 120-line file. Aionrs-specific helpers
(map_aionrs_provider, resolve_aionrs_url_and_compat, etc.) moved into
aionrs.rs.

Pure code motion — no behaviour change.
